Reorganize datakit::value to multiple files + errors
