+++ src/table.rs
@@ -1,4 +1,7 @@
-use crate::value::*;
+use crate::errors::*;
+use crate::value::constraint::*;
+use crate::value::definition::*;
+use crate::value::traits::*;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
